requirements.txt
pandas
numpy
requests
tqdm
pyyaml
python-dotenv
openai
streamlit
matplotlib
scipy
scikit-learn
weasyprint    # optional for HTML->PDF
xlrd
openpyxl
tenacity
joblib
